Refactor elapsed time types from u128 to u64 across multiple modules
- Updated `CorrelationWindows` to use u64 for time windows instead of u128.
- Changed session schema validation to access schema version from `session.core`.
- Adjusted `CompactRunSummary` and related functions to reflect changes in session structure.
- Modified task-related structures and functions to use u64 for elapsed time.
- Updated TUI rendering logic to accommodate new elapsed time type.
- Refactored comparability metrics to use u64 for elapsed time tracking.
- Adjusted warmup retention logic to work with u64 elapsed time.

+++ stutter/src/hwmon.rs
@@ -251,7 +251,7 @@ impl HwmonReader {
         reader
     }
 
-    pub fn sample(&mut self, elapsed_ms: u128) -> GpuSample {
+    pub fn sample(&mut self, elapsed_ms: u64) -> GpuSample {
         let mut gpu_busy_percent = read_u32_cached(&mut self.gpu_busy, &mut self.buf);
         let mut vram_used_bytes = read_u64_cached(&mut self.vram_used, &mut self.buf);
         let mut vram_total_bytes = read_u64_cached(&mut self.vram_total, &mut self.buf);
